+++ pmd-cpp/src/test/java/net/sourceforge/pmd/cpd/CPPTokenizerTest.java
@@ -159,6 +159,74 @@ public class CPPTokenizerTest {
         tokenizer.tokenize(code, new Tokens());
     }
 
+    public void testStringPrefix(String code, String expToken, int tokenIndex, int expNoTokens) {
+        final Tokens tokens = parse(code);
+        final TokenEntry token = tokens.getTokens().get(tokenIndex);
+        assertEquals(expNoTokens, tokens.size());
+        assertEquals(expToken, token.toString());
+    }
+
+    public void testCharacterPrefix(String code, String expToken) {
+        testStringPrefix(code, expToken, 3, 6);
+    }
+
+    public void testStringPrefix(String code, String expToken) {
+        testStringPrefix(code, expToken, 5, 8);
+    }
+
+    @Test
+    public void testCharacterPrefixNoPrefix() {
+        testCharacterPrefix("char a =  '\\x30';", "'\\x30'");
+    }
+
+    @Test
+    public void testCharacterPrefixWideCharacter() {
+        testCharacterPrefix("wchar_t b = L'\\xFFEF';", "L'\\xFFEF'");
+    }
+
+    @Test
+    public void testCharacterPrefixChar16() {
+        testCharacterPrefix("char16_t c = u'\\u00F6';", "u'\\u00F6'");
+    }
+
+    @Test
+    public void testCharacterPrefixChar32() {
+        testCharacterPrefix("char32_t d = U'\\U0010FFFF';", "U'\\U0010FFFF'");
+    }
+
+    @Test
+    public void testStringPrefixNoPrefix() {
+        testStringPrefix("char A[] = \"Hello\\x0A\";", "\"Hello\\x0A\"");
+    }
+
+    @Test
+    public void testStringPrefixWideString() {
+        testStringPrefix("wchar_t B[] = L\"Hell\\xF6\\x0A\";", "L\"Hell\\xF6\\x0A\"");
+    }
+
+    @Test
+    public void testStringPrefixChar16() {
+        testStringPrefix("char16_t C[] = u\"Hell\\u00F6\";", "u\"Hell\\u00F6\"");
+    }
+
+    @Test
+    public void testStringPrefixChar32() {
+        testStringPrefix("char32_t D[] = U\"Hell\\U000000F6\\U0010FFFF\";", "U\"Hell\\U000000F6\\U0010FFFF\"");
+    }
+
+    @Test
+    public void testStringPrefixUtf8() {
+        testStringPrefix("auto E[] = u8\"\\u00F6\\U0010FFFF\";", "u8\"\\u00F6\\U0010FFFF\"");
+    }
+
+    @Test
+    public void testRawStringLiterals() throws IOException {
+        final String code = IOUtils.toString(CPPTokenizerTest.class.getResourceAsStream("cpp/issue-1784.cpp"), StandardCharsets.UTF_8);
+        Tokens tokens = parse(code);
+        assertTrue(TokenEntry.getEOF() != tokens.getTokens().get(0));
+        assertEquals(16, tokens.size());
+    }
+
     @Test
     public void testDigitSeparators() {
         final String code = "auto integer_literal = 1'000'000;" + PMD.EOL